vibe tested on semi complete database: mostly works

+++ app/src/main/java/com/example/chadlohrli/myapplication/SharedPrefs.java
@@ -16,7 +16,7 @@ import static android.content.Context.MODE_PRIVATE;
 
 public class SharedPrefs {
 
-    public static void saveSongData(Context context, String id, float latitude, float longitude, int day, int time, float rating, int state, int timesPlayed, String lastPlayed, int fav) {
+    public static void saveSongData(Context context, String id, float latitude, float longitude, int day, int time, int rating, int state, int timesPlayed, String lastPlayed, int fav) {
 
         SharedPreferences newSong = context.getSharedPreferences(id, MODE_PRIVATE);
         SharedPreferences.Editor songadd = newSong.edit();
@@ -26,7 +26,7 @@ public class SharedPrefs {
         songadd.putFloat("Longitude", longitude);
         songadd.putInt("Day", day);
         songadd.putInt("Time", time);
-        songadd.putFloat("Rating", rating);
+        songadd.putInt("Rating", rating);
         songadd.putInt("State",state);
         songadd.putInt("Times played", timesPlayed);
         songadd.putString("Last played", lastPlayed);
@@ -64,13 +64,13 @@ public class SharedPrefs {
         songadd.commit();
     }
 
-    public static void updateRating(Context context, String id, float rating) {
+    public static void updateRating(Context context, String id, int rating) {
 
         SharedPreferences newSong = context.getSharedPreferences(id, MODE_PRIVATE);
         SharedPreferences.Editor songadd = newSong.edit();
-        songadd.putFloat("Rating", rating);
+        songadd.putInt("Rating", rating);
 
-        Log.i("Rating Updated", Float.toString(rating));
+        Log.i("Rating Updated", Integer.toString(rating));
         songadd.commit();
     }
